Move with offline check to stores utils folder outside tests
